Modify: update README.md in docs folder
