+++ build.gradle
@@ -43,7 +43,7 @@ test {
 }
 
 application {
-    mainClass.set("duke.DukeLauncher")
+    mainClass.set("zirnitra.ZirnitraLauncher")
 }
 
 shadowJar {
