+++ docs/README.md
@@ -1,30 +1,37 @@
-# Duke.Duke User Guide
+# Zirnitra Chatbot
 
-// Update the title above to match the actual product name
+> “Your mind is for having ideas, not holding them.” – David Allen ([source](https://dansilvestre.com/productivity-quotes))
 
-// Product screenshot goes here
+Zirnitra 🐉 frees your mind of having to remember things you need to do. It's,
 
-// Product intro goes here
+- text-based
+- easy to learn
+- ~~FAST~~ SUPER FAST to use
 
-## Adding deadlines
+All you need to do is:
 
-// Describe the action and its outcome.
+1. download it [here](https://github.com/mjyan02/ip.git).
+2. double-click it.
+3. add your tasks.
+4. let it manage your tasks for you 😉
 
-// Give examples of usage
+And it is **FREE**!
 
-Example: `keyword (optional arguments)`
+**Features:**
 
-// A description of the expected outcome goes here
+- [x] Add new tasks using **"todo", "deadline"** and **"event"** keywords
+- [x] **List, mark, unmark** or **delete** any tasks added
+- [x] Locate specific tasks in the list using **"find"** keyword
+- [x] Duplicate tasks are detected and prevented from being added
+
+---
+
+> [!NOTE] If you are a Java programmer, you can use it to practice Java too. Here's the `main` method:
 
 ```
-expected output
+public class Main {
+    public static void main(String[] args) {
+        Application.launch(DukeLauncher.class, args);
+    }
+}
 ```
-
-## Feature ABC
-
-// Feature details
-
-
-## Feature XYZ
-
-// Feature details
