--- src/main/java/Duke/DukeLauncher.java
@@ -1,14 +1,0 @@
-package duke;
-
-import duke.ui.DukeGui;
-
-import javafx.application.Application;
-
-/**
- * Launcher for Duke GUI.
- */
-public class DukeLauncher {
-    public static void main(String[] args) {
-        Application.launch(DukeGui.class, args);
-    }
-}
