+++ src/main/java/zirnitra/ZirnitraLauncher.java
@@ -1,0 +1,14 @@
+package zirnitra;
+
+import zirnitra.ui.ZirnitraGui;
+
+import javafx.application.Application;
+
+/**
+ * Launcher for Zirnitra GUI.
+ */
+public class ZirnitraLauncher {
+    public static void main(String[] args) {
+        Application.launch(ZirnitraGui.class, args);
+    }
+}
