+++ src/main/java/zirnitra/task/Event.java
@@ -1,0 +1,82 @@
+package zirnitra.task;
+
+import zirnitra.exceptions.ZirnitraException;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
+/**
+ * Represents an event task with a description, start date and end date.
+ */
+public class Event extends Task {
+    /**
+     * The start date of the event.
+     */
+    protected LocalDate from;
+
+    /**
+     * The end date of the event.
+     */
+    protected LocalDate to;
+
+    /**
+     * Constructs a new Event task with the given description, start date and end date.
+     *
+     * @param description The description of the event.
+     * @param from The start date of the event in yyyy-MM-dd format.
+     * @param to The end date of the event in yyyy-MM-dd format.
+     * @throws ZirnitraException If the date format is invalid.
+     */
+    public Event(String description, String from, String to) throws ZirnitraException {
+        super(description);
+
+        try {
+            this.from = LocalDate.parse(from);
+            this.to = LocalDate.parse(to);
+        } catch (DateTimeParseException e) {
+            throw new ZirnitraException("Invalid date format! Please use yyyy-MM-dd!");
+        }
+    }
+
+    /**
+     * Constructs a new Event task with the given description, start date and end date.
+     *
+     * @param description The description of the event.
+     * @param from The start date of the event in yyyy-MM-dd format.
+     * @param to The end date of the event in yyyy-MM-dd format.
+     * @param isDone Specifies if task is marked
+     * @throws ZirnitraException If the date format is invalid.
+     */
+    public Event(String description, String from, String to, boolean isDone) throws ZirnitraException {
+        super(description, isDone);
+
+        try {
+            this.from = LocalDate.parse(from);
+            this.to = LocalDate.parse(to);
+        } catch (DateTimeParseException e) {
+            throw new ZirnitraException("Invalid date format! Please use yyyy-MM-dd!");
+        }
+    }
+
+    /**
+     * Converts the task into a formatted string representation suitable for file storage on hard disk.
+     *
+     * @return A string representation of the task for file storage.
+     */
+    @Override
+    public String toFile() {
+        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from + " | " + to;
+    }
+
+    /**
+     * Returns a string representation of the task.
+     *
+     * @return The formatted string representation of the task.
+     */
+    @Override
+    public String toString() {
+        return "[E]" + super.toString() + " (from: " + from.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) +
+                " to: " + to.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
+    }
+}
