--- src/test/java/Duke/Exceptions/DukeExceptionTest.java
@@ -1,4 +1,0 @@
-package duke.exceptions;
-
-public class DukeExceptionTest {
-}
