--- src/test/java/Duke/ParserTest.java
@@ -1,4 +1,0 @@
-package duke;
-
-public class ParserTest {
-}
