--- src/test/java/Duke/StorageTest.java
@@ -1,4 +1,0 @@
-package duke;
-
-public class StorageTest {
-}
