--- src/test/java/Duke/Task/KeywordsTest.java
@@ -1,4 +1,0 @@
-package duke.task;
-
-public class KeywordsTest {
-}
