--- src/test/java/Duke/Task/TaskTest.java
@@ -1,4 +1,0 @@
-package duke.task;
-
-public class TaskTest {
-}
