--- src/test/java/Duke/TaskListTest.java
@@ -1,4 +1,0 @@
-package duke;
-
-public class TaskListTest {
-}
