--- src/test/java/Duke/Ui/UiTest.java
@@ -1,4 +1,0 @@
-package duke.ui;
-
-public class UiTest {
-}
