+++ src/test/java/zirnitra/ParserTest.java
@@ -1,0 +1,4 @@
+package zirnitra;
+
+public class ParserTest {
+}
