+++ src/test/java/zirnitra/StorageTest.java
@@ -1,0 +1,4 @@
+package zirnitra;
+
+public class StorageTest {
+}
