+++ src/test/java/zirnitra/exceptions/ZirnitraExceptionTest.java
@@ -1,0 +1,4 @@
+package zirnitra.exceptions;
+
+public class ZirnitraExceptionTest {
+}
