+++ src/test/java/zirnitra/task/KeywordsTest.java
@@ -1,0 +1,4 @@
+package zirnitra.task;
+
+public class KeywordsTest {
+}
