+++ src/test/java/zirnitra/task/TaskTest.java
@@ -1,0 +1,4 @@
+package zirnitra.task;
+
+public class TaskTest {
+}
