+++ src/test/java/zirnitra/ui/UiTest.java
@@ -1,0 +1,4 @@
+package zirnitra.ui;
+
+public class UiTest {
+}
